+++ src/smsfusion/__init__.py
@@ -1,13 +1,11 @@
+from . import noise
 from ._ahrs import AHRS
 from ._ins import AidedINS, StrapdownINS, gravity
-from ._noise_models import gauss_markov, random_walk, white_noise
 
 __all__ = [
     "AHRS",
     "AidedINS",
-    "gauss_markov",
     "gravity",
-    "random_walk",
+    "noise",
     "StrapdownINS",
-    "white_noise",
 ]
