+++ tests/test_noise_models.py
@@ -4,8 +4,8 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from smsfusion import gauss_markov, random_walk, white_noise
-from smsfusion._noise_models import _standard_normal
+from smsfusion.noise import gauss_markov, random_walk, white_noise
+from smsfusion.noise._noise_models import _standard_normal
 
 TEST_PATH = Path(__file__).parent
